Rename WCPaymentAccountResult to PaymentAccountResult
